refactor(*): MVC 패턴 적용을 위해 'BaseballGame', 'Controller' 클래스 분리
