+++ docs/README.md
@@ -79,8 +79,9 @@
 [x] 랜덤 숫자 생성 로직 `BaseballNumberMaker` 파일로 분리  
 [x] `Validator` 파일로 예외처리 로직 분리    
 [ ] 함수 라인을 15라인으로 제한하면서 한가지 기능만을 하도록 함    
-[ ] MVC패턴을 적용하여 객체 분리     
+[x] MVC패턴을 적용하여 객체 분리     
 [ ] Airbnb JS Style Guide에 따라 코드 작성
+[ ] else 지양
 
 # ⚾️ 숫자 야구 App 규칙
 - <U>1부터 9까지의 서로 다른 수</U>로 이루어진 3자리의 수를 맞추는 게임
