+++ src/BaseballGame.js
@@ -1,0 +1,16 @@
+const { createRandomNumber } =require('./BaseballNumberMaker');
+
+class BaseballGame {
+
+    #randomNumber;
+
+    setRandomNumber() {
+        this.#randomNumber = createRandomNumber();
+      }
+    
+    getRandomNumber() {
+        return this.#randomNumber;
+      }
+}
+
+module.exports = BaseballGame;
